+++ temba/channels/tests.py
@@ -1581,46 +1581,69 @@ class ChannelLogTest(TembaTest):
         )
         msg_out = self.create_outgoing_msg(contact, "Working", channel=channel, status="S", logs=[log])
 
-        expected_unredacted = {
-            "description": "Message Send",
-            "http_logs": [
-                {
-                    "url": "https://telegram.com/send?to=74747474",
-                    "status_code": 400,
-                    "request": 'POST https://telegram.com/send?to=74747474 HTTP/1.1\r\n\r\n{"to":"74747474"}',
-                    "response": 'HTTP/2.0 200 OK\r\n\r\n{"to":"74747474","first_name":"Fred"}',
-                    "elapsed_ms": 263,
-                    "retries": 0,
-                    "created_on": "2022-08-17T14:07:30Z",
-                }
-            ],
-            "errors": [{"code": "bad_response", "ext_code": "", "message": "response not right", "ref_url": None}],
-            "created_on": matchers.Datetime(),
-        }
+        self.assertEqual(
+            {
+                "type": "msg_send",
+                "http_logs": [
+                    {
+                        "url": "https://telegram.com/send?to=74747474",
+                        "status_code": 400,
+                        "request": 'POST https://telegram.com/send?to=74747474 HTTP/1.1\r\n\r\n{"to":"74747474"}',
+                        "response": 'HTTP/2.0 200 OK\r\n\r\n{"to":"74747474","first_name":"Fred"}',
+                        "elapsed_ms": 263,
+                        "retries": 0,
+                        "created_on": "2022-08-17T14:07:30Z",
+                    }
+                ],
+                "errors": [{"code": "bad_response", "ext_code": "", "message": "response not right", "ref_url": None}],
+                "elapsed_ms": 0,
+                "created_on": matchers.ISODate(),
+            },
+            log.get_display(anonymize=False, urn=msg_out.contact_urn),
+        )
 
-        expected_redacted = {
-            "description": "Message Send",
-            "http_logs": [
-                {
-                    "url": "https://telegram.com/send?to=********",
-                    "status_code": 400,
-                    "request": 'POST https://telegram.com/send?to=******** HTTP/1.1\r\n\r\n{"to":"********"}',
-                    "response": 'HTTP/2.0 200 OK\r\n\r\n{"to": "********", "first_name": "********"}',
-                    "elapsed_ms": 263,
-                    "retries": 0,
-                    "created_on": "2022-08-17T14:07:30Z",
-                }
-            ],
-            "errors": [{"code": "bad_response", "ext_code": "", "message": "response n********", "ref_url": None}],
-            "created_on": matchers.Datetime(),
-        }
+        self.assertEqual(
+            {
+                "type": "msg_send",
+                "http_logs": [
+                    {
+                        "url": "https://telegram.com/send?to=********",
+                        "status_code": 400,
+                        "request": 'POST https://telegram.com/send?to=******** HTTP/1.1\r\n\r\n{"to":"********"}',
+                        "response": 'HTTP/2.0 200 OK\r\n\r\n{"to": "********", "first_name": "********"}',
+                        "elapsed_ms": 263,
+                        "retries": 0,
+                        "created_on": "2022-08-17T14:07:30Z",
+                    }
+                ],
+                "errors": [{"code": "bad_response", "ext_code": "", "message": "response n********", "ref_url": None}],
+                "elapsed_ms": 0,
+                "created_on": matchers.ISODate(),
+            },
+            log.get_display(anonymize=True, urn=msg_out.contact_urn),
+        )
 
-        self.assertEqual(expected_unredacted, log.get_display(self.admin, urn=msg_out.contact_urn))
-        self.assertEqual(expected_unredacted, log.get_display(self.customer_support, urn=msg_out.contact_urn))
-
-        with AnonymousOrg(self.org):
-            self.assertEqual(expected_redacted, log.get_display(self.admin, urn=msg_out.contact_urn))
-            self.assertEqual(expected_unredacted, log.get_display(self.customer_support, urn=msg_out.contact_urn))
+        # if we don't pass it a URN, anonymization is more aggressive
+        self.assertEqual(
+            {
+                "type": "msg_send",
+                "http_logs": [
+                    {
+                        "url": "https://te********",
+                        "status_code": 400,
+                        "request": "POST https********",
+                        "response": "HTTP/2.0 2********",
+                        "elapsed_ms": 263,
+                        "retries": 0,
+                        "created_on": "2022-08-17T14:07:30Z",
+                    }
+                ],
+                "errors": [{"code": "bad_response", "ext_code": "", "message": "response n********", "ref_url": None}],
+                "elapsed_ms": 0,
+                "created_on": matchers.ISODate(),
+            },
+            log.get_display(anonymize=True, urn=None),
+        )
 
     def test_get_display_timed_out(self):
         channel = self.create_channel("TG", "Telegram", "mybot")
@@ -1642,46 +1665,43 @@ class ChannelLogTest(TembaTest):
         )
         msg_out = self.create_outgoing_msg(contact, "Working", channel=channel, status="S", logs=[log])
 
-        expected_unredacted = {
-            "description": "Message Send",
-            "http_logs": [
-                {
-                    "url": "https://telegram.com/send?to=74747474",
-                    "status_code": 0,
-                    "request": 'POST https://telegram.com/send?to=74747474 HTTP/1.1\r\n\r\n{"to":"74747474"}',
-                    "response": "",
-                    "elapsed_ms": 30001,
-                    "retries": 0,
-                    "created_on": "2022-08-17T14:07:30Z",
-                }
-            ],
-            "errors": [{"code": "bad_response", "ext_code": "", "message": "response not right", "ref_url": None}],
-            "created_on": matchers.Datetime(),
-        }
-
-        expected_redacted = {
-            "description": "Message Send",
-            "http_logs": [
-                {
-                    "url": "https://telegram.com/send?to=********",
-                    "status_code": 0,
-                    "request": 'POST https://telegram.com/send?to=******** HTTP/1.1\r\n\r\n{"to":"********"}',
-                    "response": "********",
-                    "elapsed_ms": 30001,
-                    "retries": 0,
-                    "created_on": "2022-08-17T14:07:30Z",
-                }
-            ],
-            "errors": [{"code": "bad_response", "ext_code": "", "message": "response n********", "ref_url": None}],
-            "created_on": matchers.Datetime(),
-        }
-
-        self.assertEqual(expected_unredacted, log.get_display(self.admin, urn=msg_out.contact_urn))
-        self.assertEqual(expected_unredacted, log.get_display(self.customer_support, urn=msg_out.contact_urn))
-
-        with AnonymousOrg(self.org):
-            self.assertEqual(expected_redacted, log.get_display(self.admin, urn=msg_out.contact_urn))
-            self.assertEqual(expected_unredacted, log.get_display(self.customer_support, urn=msg_out.contact_urn))
+        self.assertEqual(
+            {
+                "type": "msg_send",
+                "http_logs": [
+                    {
+                        "url": "https://telegram.com/send?to=74747474",
+                        "request": 'POST https://telegram.com/send?to=74747474 HTTP/1.1\r\n\r\n{"to":"74747474"}',
+                        "elapsed_ms": 30001,
+                        "retries": 0,
+                        "created_on": "2022-08-17T14:07:30Z",
+                    }
+                ],
+                "errors": [{"code": "bad_response", "ext_code": "", "message": "response not right", "ref_url": None}],
+                "elapsed_ms": 0,
+                "created_on": matchers.ISODate(),
+            },
+            log.get_display(anonymize=False, urn=msg_out.contact_urn),
+        )
+        self.assertEqual(
+            {
+                "type": "msg_send",
+                "http_logs": [
+                    {
+                        "url": "https://telegram.com/send?to=********",
+                        "request": 'POST https://telegram.com/send?to=******** HTTP/1.1\r\n\r\n{"to":"********"}',
+                        "response": "",
+                        "elapsed_ms": 30001,
+                        "retries": 0,
+                        "created_on": "2022-08-17T14:07:30Z",
+                    }
+                ],
+                "errors": [{"code": "bad_response", "ext_code": "", "message": "response n********", "ref_url": None}],
+                "elapsed_ms": 0,
+                "created_on": matchers.ISODate(),
+            },
+            log.get_display(anonymize=True, urn=msg_out.contact_urn),
+        )
 
     def test_trim_task(self):
         ChannelLog.objects.create(
@@ -1900,6 +1920,16 @@ class ChannelLogCRUDLTest(CRUDLTestMixin, TembaTest):
         response = self.client.get(reverse("channels.channellog_list", args=["invalid-uuid"]))
         self.assertEqual(404, response.status_code)
 
+    def assertRedacted(self, response, values: tuple):
+        for value in values:
+            self.assertNotContains(response, value)
+
+        self.assertContains(response, ChannelLog.REDACT_MASK)
+
+    def assertNotRedacted(self, response, values: tuple):
+        for value in values:
+            self.assertContains(response, value)
+
     def test_redaction_for_telegram(self):
         urn = "telegram:3527065"
         contact = self.create_contact("Fred Jones", urns=[urn])
@@ -1928,35 +1958,22 @@ class ChannelLogCRUDLTest(CRUDLTestMixin, TembaTest):
 
         # check read page shows un-redacted content for a regular org
         response = self.client.get(read_url)
+        self.assertNotRedacted(response, ("3527065", "Nic", "Pottier"))
 
-        self.assertContains(response, "3527065", count=3)
-        self.assertContains(response, "Nic", count=2)
-        self.assertContains(response, "Pottier", count=1)
-
-        # check read page shows redacted content for an anon org
+        # but for anon org we see redaction...
         with AnonymousOrg(self.org):
             response = self.client.get(read_url)
+            self.assertRedacted(response, ("3527065", "Nic", "Pottier"))
 
-            self.assertContains(response, "3527065", count=0)
-            self.assertContains(response, "Nic", count=0)
-            self.assertContains(response, "Pottier", count=0)
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=8)
+            # even as customer support
+            self.login(self.customer_support, choose_org=self.org)
 
-        # login as customer support, must see URNs
-        self.login(self.customer_support, choose_org=self.org)
-
-        response = self.client.get(read_url)
-
-        self.assertContains(response, "3527065", count=3)
-
-        with AnonymousOrg(self.org):
             response = self.client.get(read_url)
-            # contact_urn is still masked on the read page, it uses contacts.models.Contact.get_display
-            # Contact.get_display does not check if user is staff
-            self.assertContains(response, "3527065", count=2)
-            self.assertContains(response, "Nic", count=2)
-            self.assertContains(response, "Pottier", count=1)
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=1)
+            self.assertRedacted(response, ("3527065", "Nic", "Pottier"))
+
+            # unless we explicitly break out of it
+            response = self.client.get(read_url + "?break=1")
+            self.assertNotRedacted(response, ("3527065", "Nic", "Pottier"))
 
     def test_redaction_for_telegram_with_invalid_json(self):
         urn = "telegram:3527065"
@@ -1983,32 +2000,15 @@ class ChannelLogCRUDLTest(CRUDLTestMixin, TembaTest):
         self.login(self.admin)
 
         read_url = reverse("channels.channellog_msg", args=[channel.uuid, msg.id])
+
+        # check read page shows un-redacted content for a regular org
         response = self.client.get(read_url)
+        self.assertNotRedacted(response, ("3527065", "Nic"))
 
-        self.assertContains(response, "3527065", count=2)
-
+        # but for anon org we see redaction...
         with AnonymousOrg(self.org):
             response = self.client.get(read_url)
-
-            self.assertContains(response, "3527065", count=0)
-            self.assertContains(response, "Nic", count=0)
-            self.assertContains(response, "Pottier", count=0)
-
-            # everything is masked
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=3)
-
-        # login as customer support, must see URNs
-        self.login(self.customer_support, choose_org=self.org)
-
-        response = self.client.get(read_url)
-
-        self.assertContains(response, "3527065", count=2)
-
-        with AnonymousOrg(self.org):
-            response = self.client.get(read_url)
-            self.assertContains(response, "Nic", count=1)
-
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=1)
+            self.assertRedacted(response, ("3527065", "Nic"))
 
     def test_redaction_for_telegram_when_no_match(self):
         urn = "telegram:3527065"
@@ -2035,32 +2035,15 @@ class ChannelLogCRUDLTest(CRUDLTestMixin, TembaTest):
         self.login(self.admin)
 
         read_url = reverse("channels.channellog_msg", args=[channel.uuid, msg.id])
+
+        # check read page shows un-redacted content for a regular org
         response = self.client.get(read_url)
+        self.assertNotRedacted(response, ("3527065",))
 
-        self.assertContains(response, "3527065", count=1)
-        self.assertContains(response, "There is no contact identifying information", count=2)
-
+        # but for anon org we see complete redaction
         with AnonymousOrg(self.org):
             response = self.client.get(read_url)
-
-            # url/request/reponse are masked
-            self.assertContains(response, "There is no contact identifying information", count=0)
-
-            self.assertContains(response, "3527065", count=0)
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=3)
-
-        # login as customer support, must see URNs
-        self.login(self.customer_support, choose_org=self.org)
-
-        response = self.client.get(read_url)
-
-        self.assertContains(response, "3527065", count=1)
-
-        with AnonymousOrg(self.org):
-            response = self.client.get(read_url)
-            self.assertContains(response, "There is no contact identifying information", count=2)
-
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=1)
+            self.assertRedacted(response, ("3527065", "api.telegram.org", "/65474/sendMessage"))
 
     def test_redaction_for_twitter(self):
         urn = "twitterid:767659860"
@@ -2088,36 +2071,14 @@ class ChannelLogCRUDLTest(CRUDLTestMixin, TembaTest):
 
         read_url = reverse("channels.channellog_msg", args=[channel.uuid, msg.id])
 
+        # check read page shows un-redacted content for a regular org
         response = self.client.get(read_url)
+        self.assertNotRedacted(response, ("767659860", "Aaron Tumukunde", "tumaaron"))
 
-        self.assertContains(response, "767659860", count=5)
-        self.assertContains(response, "Aaron Tumukunde", count=1)
-        self.assertContains(response, "tumaaron", count=2)
-
+        # but for anon org we see redaction...
         with AnonymousOrg(self.org):
             response = self.client.get(read_url)
-
-            self.assertContains(response, "767659860", count=0)
-            self.assertContains(response, "Aaron Tumukunde", count=0)
-            self.assertContains(response, "tumaaron", count=0)
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=13)
-
-        # login as customer support, must see URNs
-        self.login(self.customer_support, choose_org=self.org)
-
-        response = self.client.get(read_url)
-
-        self.assertContains(response, "767659860", count=5)
-
-        with AnonymousOrg(self.org):
-            response = self.client.get(read_url)
-            # contact_urn is still masked on the read page, it uses contacts.models.Contact.get_display
-            # Contact.get_display does not check if user is staff
-            self.assertContains(response, "767659860", count=4)
-            self.assertContains(response, "Aaron Tumukunde", count=1)
-            self.assertContains(response, "tumaaron", count=2)
-
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=1)
+            self.assertRedacted(response, ("767659860", "Aaron Tumukunde", "tumaaron"))
 
     def test_redaction_for_twitter_when_no_match(self):
         urn = "twitterid:767659860"
@@ -2144,31 +2105,15 @@ class ChannelLogCRUDLTest(CRUDLTestMixin, TembaTest):
         self.login(self.admin)
 
         read_url = reverse("channels.channellog_msg", args=[channel.uuid, msg.id])
+
+        # check read page shows un-redacted content for a regular org
         response = self.client.get(read_url)
+        self.assertNotRedacted(response, ("767659860",))
 
-        self.assertContains(response, "767659860", count=1)
-        self.assertContains(response, "There is no contact identifying information", count=2)
-
+        # but for anon org we see complete redaction...
         with AnonymousOrg(self.org):
             response = self.client.get(read_url)
-
-            # url/request/reponse are masked
-            self.assertContains(response, "There is no contact identifying information", count=0)
-
-            self.assertContains(response, "767659860", count=0)
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=3)
-
-        # login as customer support, must see URNs
-        self.login(self.customer_support, choose_org=self.org)
-        response = self.client.get(read_url)
-
-        self.assertContains(response, "767659860", count=1)
-
-        with AnonymousOrg(self.org):
-            response = self.client.get(read_url)
-            self.assertContains(response, "There is no contact identifying information", count=2)
-
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=1)
+            self.assertRedacted(response, ("767659860", "twitter.com", "/65474/sendMessage"))
 
     def test_redaction_for_facebook(self):
         urn = "facebook:2150393045080607"
@@ -2198,35 +2143,14 @@ class ChannelLogCRUDLTest(CRUDLTestMixin, TembaTest):
 
         response = self.client.get(read_url)
 
-        self.assertContains(response, "2150393045080607", count=3)
-        self.assertContains(response, "facebook:2150393045080607", count=1)
-
-        with AnonymousOrg(self.org):
-            response = self.client.get(read_url)
-
-            self.assertContains(response, "2150393045080607", count=0)
-            self.assertContains(response, "facebook:", count=1)
-
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=3)
-
-        # login as customer support, must see URNs
-        self.login(self.customer_support, choose_org=self.org)
-
-        read_url = reverse("channels.channellog_msg", args=[channel.uuid, msg.id])
-
+        # check read page shows un-redacted content for a regular org
         response = self.client.get(read_url)
+        self.assertNotRedacted(response, ("2150393045080607",))
 
-        self.assertContains(response, "2150393045080607", count=3)
-        self.assertContains(response, "facebook:2150393045080607", count=1)
-
+        # but for anon org we see redaction...
         with AnonymousOrg(self.org):
             response = self.client.get(read_url)
-            # contact_urn is still masked on the read page, it uses contacts.models.Contact.get_display
-            # Contact.get_display does not check if user is staff
-            self.assertContains(response, "2150393045080607", count=2)
-            self.assertContains(response, "facebook:", count=1)
-
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=1)
+            self.assertRedacted(response, ("2150393045080607",))
 
     def test_redaction_for_facebook_when_no_match(self):
         # in this case we are paranoid and mask everything
@@ -2255,33 +2179,14 @@ class ChannelLogCRUDLTest(CRUDLTestMixin, TembaTest):
 
         read_url = reverse("channels.channellog_msg", args=[channel.uuid, msg.id])
 
+        # check read page shows un-redacted content for a regular org
         response = self.client.get(read_url)
+        self.assertNotRedacted(response, ("2150393045080607",))
 
-        self.assertContains(response, "2150393045080607", count=1)
-
+        # but for anon org we see complete redaction...
         with AnonymousOrg(self.org):
             response = self.client.get(read_url)
-
-            # url/request/reponse are masked
-            self.assertContains(response, "There is no contact identifying information", count=0)
-
-            self.assertContains(response, "2150393045080607", count=0)
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=3)
-
-        # login as customer support, must see URNs
-        self.login(self.customer_support, choose_org=self.org)
-
-        response = self.client.get(read_url)
-
-        self.assertContains(response, "2150393045080607", count=1)
-
-        with AnonymousOrg(self.org):
-            response = self.client.get(read_url)
-            self.assertContains(response, "There is no contact identifying information", count=2)
-
-            # contact_urn is still masked on the read page, it uses contacts.models.Contact.get_display
-            # Contact.get_display does not check if user is staff
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=1)
+            self.assertRedacted(response, ("2150393045080607", "facebook.com", "/65474/sendMessage"))
 
     def test_redaction_for_twilio(self):
         contact = self.create_contact("Fred Jones", phone="+593979099111")
@@ -2310,43 +2215,18 @@ class ChannelLogCRUDLTest(CRUDLTestMixin, TembaTest):
 
         # check read page shows un-redacted content for a regular org
         response = self.client.get(read_url)
+        self.assertNotRedacted(response, ("097 909 9111", "979099111", "Quito"))
 
-        self.assertContains(response, "097 909 9111", count=1)
-        self.assertContains(response, "979099111", count=1)
-        self.assertContains(response, "Quito", count=1)
-
-        # check read page shows redacted content for an anon org
+        # but for anon org we see redaction...
         with AnonymousOrg(self.org):
             response = self.client.get(read_url)
+            self.assertRedacted(response, ("097 909 9111", "979099111", "Quito"))
 
-            self.assertContains(response, "097 909 9111", count=0)
-            self.assertContains(response, "979099111", count=0)
-            self.assertContains(response, "Quito", count=0)
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=4)
-
-        # login as customer support, must see URNs
-        self.login(self.customer_support, choose_org=self.org)
-
-        response = self.client.get(read_url)
-
-        self.assertContains(response, "097 909 9111", count=1)
-        self.assertContains(response, "979099111", count=1)
-        self.assertContains(response, "Quito", count=1)
-
-        with AnonymousOrg(self.org):
-            response = self.client.get(read_url)
-            # contact_urn is still masked on the read page, it uses contacts.models.Contact.get_display
-            # Contact.get_display does not check if user is staff
-            self.assertContains(response, "097 909 9111", count=0)
-            self.assertContains(response, "979099111", count=1)
-            self.assertContains(response, "Quito", count=1)
-            self.assertContains(response, HTTPLog.REDACT_MASK, count=1)
-
-    def test_channellog_hide_whatsapp_cloud(self):
+    def test_channellog_whatsapp_cloud(self):
         urn = "whatsapp:15128505839"
         contact = self.create_contact("Fred Jones", urns=[urn])
         channel = self.create_channel("WAC", "Test WAC Channel", "54764868534")
-        success_log = ChannelLog.objects.create(
+        log = ChannelLog.objects.create(
             channel=channel,
             log_type=ChannelLog.LOG_TYPE_MSG_SEND,
             is_error=False,
@@ -2363,7 +2243,6 @@ Content-Length: 343
 Content-Type: application/x-www-form-urlencoded
 User-Agent: SignalwireCallback/1.0
 Authorization: Bearer {settings.WHATSAPP_ADMIN_SYSTEM_USER_TOKEN}
-
 MessageSid=e1d12194-a643-4007-834a-5900db47e262&SmsSid=e1d12194-a643-4007-834a-5900db47e262&AccountSid=<redacted>&From=%2B15618981512&To=%2B15128505839&Body=Hi+Ben+Google+Voice%2C+Did+you+enjoy+your+stay+at+White+Bay+Villas%3F++Answer+with+Yes+or+No.+reply+STOP+to+opt-out.&NumMedia=0&NumSegments=1&MessageStatus=sent""",
                     "response": '{"success": true }',
                     "elapsed_ms": 12,
@@ -2372,16 +2251,15 @@ MessageSid=e1d12194-a643-4007-834a-5900db47e262&SmsSid=e1d12194-a643-4007-834a-5
                 }
             ],
         )
-        self.create_incoming_msg(contact, "incoming msg", channel=channel, logs=[success_log])
+        self.create_incoming_msg(contact, "incoming msg", channel=channel, logs=[log])
 
         self.login(self.admin)
 
-        read_url = reverse("channels.channellog_read", args=[success_log.id])
+        read_url = reverse("channels.channellog_read", args=[log.id])
 
-        response = self.client.get(read_url)
-        self.assertNotContains(response, settings.WHATSAPP_ADMIN_SYSTEM_USER_TOKEN)
-        self.assertContains(response, f"/send/message?access_token={HTTPLog.REDACT_MASK}")
-        self.assertContains(response, f"Authorization: Bearer {HTTPLog.REDACT_MASK}")
+        # the token should have been redacted by courier so blow up rather than let user see it
+        with self.assertRaises(AssertionError):
+            self.client.get(read_url)
 
     def test_channellog_anonymous_org_no_msg(self):
         tw_urn = "15128505839"