+++ temba/msgs/models.py
@@ -183,7 +183,7 @@ class Broadcast(models.Model):
     STATUS_FAILED = "F"
     STATUS_CHOICES = ((STATUS_QUEUED, "Queued"), (STATUS_SENT, "Sent"), (STATUS_FAILED, "Failed"))
 
-    org = models.ForeignKey(Org, on_delete=models.PROTECT)
+    org = models.ForeignKey(Org, on_delete=models.PROTECT, related_name="broadcasts")
 
     # recipients of this broadcast
     groups = models.ManyToManyField(ContactGroup, related_name="addressed_broadcasts")
@@ -256,6 +256,13 @@ class Broadcast(models.Model):
         broadcast._set_recipients(groups=groups, contacts=contacts, urns=urns, contact_ids=contact_ids)
 
         return broadcast
+
+    @classmethod
+    def get_queued(cls, org):
+        """
+        Gets the queued broadcasts which will be prepended to the Outbox
+        """
+        return org.broadcasts.filter(status=cls.STATUS_QUEUED, schedule=None, is_active=True)
 
     @classmethod
     def preview(cls, org, *, include: mailroom.Inclusions, exclude: mailroom.Exclusions) -> tuple[str, int]:
@@ -372,7 +379,7 @@ class Broadcast(models.Model):
                 fields=["org", "-created_on"],
                 condition=Q(schedule__isnull=False, is_active=True),
             ),
-            # used to fetch pending broadcasts for the Outbox
+            # used to fetch queued broadcasts for the Outbox
             models.Index(
                 name="msgs_broadcasts_queued",
                 fields=["org", "-created_on"],